Minor improvements:
* Add hints to EditTexts
* Accept float input
* Hide keyboard after plotting

+++ app/src/main/python/plot.py
@@ -2,8 +2,8 @@ import io
 import matplotlib.pyplot as plt
 
 def plot(x, y):
-    xa = [int(word) for word in x.split()]
-    ya = [int(word) for word in y.split()]
+    xa = [float(word) for word in x.split()]
+    ya = [float(word) for word in y.split()]
 
     fig, ax = plt.subplots()
     ax.plot(xa, ya)
